apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.13
psycopg-pool==3.2.3
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
PyJWT==2.9.0
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
httpx==0.27.2
jsonschema==4.23.0
python-multipart==0.0.9
requests==2.32.3
PyYAML==6.0.2
cryptography==43.0.1
prometheus-fastapi-instrumentator==7.0.0
prometheus-client==0.21.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-httpx==0.48b0
temporalio==1.8.0
setuptools==79.0.1
